+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project are documented in this file.
 
 - Use fetch from Tauri to bypass CORS issues.
 
+## Changed
+
+- Automatically scan for compatible local servers on entering API settings
+
 ## [v0.2.2] - 2025-10-06
 
 ### Changed